+++ aiobungie/crate/components.py
@@ -29,8 +29,6 @@ since it depends on components passed to the request or due to privacy by the pr
 from __future__ import annotations
 
 __all__: tuple[str, ...] = (
-    "ComponentPrivacy",
-    "ComponentFields",
     "Component",
     "CharacterComponent",
     "ProfileComponent",
@@ -49,8 +47,9 @@ if typing.TYPE_CHECKING:
     import collections.abc as collections
 
     from aiobungie.crate import activity
-    from aiobungie.crate import items
     from aiobungie.crate import character as character_
+    from aiobungie.crate import fireteams
+    from aiobungie.crate import items
     from aiobungie.crate import profile
     from aiobungie.crate import records as records_
 
@@ -74,7 +73,7 @@ class ComponentFields(enums.Enum):
 
 # Main component cannot inherit from multiple classes that have `__slots__`
 # Which's why some components have no slots.
-@attrs.define(kw_only=True, weakref_slot=False, slots=False)
+@attrs.define(kw_only=True, slots=False)
 class RecordsComponent:
     """Represents records-only Bungie component.
 
@@ -114,7 +113,7 @@ class RecordsComponent:
     """
 
 
-@attrs.define(kw_only=True, weakref_slot=False)
+@attrs.define(kw_only=True)
 class ProfileComponent:
     """Represents a profile-only Bungie component.
 
@@ -168,51 +167,137 @@ class ProfileComponent:
     """
 
 
-@attrs.define(kw_only=True, weakref_slot=False, slots=False)
-class ItemsComponent:
-    """Represents items-only Bungie component.
+@attrs.define(kw_only=True)
+class UninstancedItemsComponent:
+    """Represents Components belonging to the player's uninstanced items."""
 
-    This includes any item related object.
-    """
-
-    item_objectives: typing.Optional[
+    objectives: typing.Optional[
         collections.Mapping[int, collections.Sequence[records_.Objective]]
     ] = attrs.field()
-    """A mapping from the character id to a sequence of its items objectives component."""
+    """A mapping from the objective id to a sequence of objectives component."""
 
-    profile_plugsets: typing.Optional[collections.Sequence[profile.ProfilePlug]] = attrs.field()
-    """A sequence of the profile's plug sets."""
+    perks: typing.Optional[
+        collections.Mapping[int, collections.Collection[items.ItemPerk]]
+    ] = attrs.field()
+    """A mapping for the item instance id to its perks."""
 
-    character_plugsets: typing.Optional[collections.Mapping[int, collections.Sequence[profile.ProfilePlug]]] = attrs.field()
-    """A mapping from the character's id to a sequence of plug objects bound to that character."""
 
-    item_instances: typing.Optional[collections.Mapping[int, ...]] = attrs.field()
+@attrs.define(kw_only=True)
+class ItemsComponent(UninstancedItemsComponent):
+    """Represents items-only Bungie component.
 
-    item_render_data: typing.Optional[collections.Mapping[int, ...]] = attrs.field()
+    This component implements most of the `ItemX` components, i.e. ItemInstances, ItemStats, ItemSockets, etc.
 
-    item_stats: typing.Optional[collections.Mapping[int, dict[str, int]]] = attrs.field()
+    Note
+    -----
+    All fields will always be `None` until either `aiobungie.ComponentType.CHARACTER_INVENTORY`
+    or `aiobungie.ComponentType.CHARACTER_EQUIPMENT` is passed to the request.
+    """
 
-    item_sockets: typing.Optional[collections.Mapping[int, collections.Sequence[profile.ProfilePlug]]] = attrs.field()
+    instances: typing.Optional[
+        collections.Sequence[collections.Mapping[int, items.ItemInstance]]
+    ] = attrs.field()
+    """A sequence from the item instance id to an item object bound to this instance.
 
-    reusable_plugs: typing.Optional[collections.Mapping[int, collections.Sequence[profile.ProfilePlug]]] = attrs.field()
+    This will be available when `aiobungie.ComponentType.ITEM_INSTANCES` is passed to the request.
+    otherwise will be `None`.
+    """
 
-    plug_objectives: typing.Optional[collections.Mapping[int, collections.Mapping[int, collections.Collection[records_.Objective]]]] = attrs.field()
+    render_data: typing.Optional[
+        collections.Mapping[int, tuple[bool, dict[int, int]]]
+    ] = attrs.field()
+    """A mapping from the item instance id to tuple that holds two values.
+
+    * First one is a bool that determines whether this item uses custom dyes or not.
+    * Second one is dict that holds int key that mapps to int value of the art regions.
+
+    This will be available when `aiobungie.ComponentType.ITEM_RENDER_DATA` is passed to the request.
+    otherwise will be `None`.
+    """
+
+    stats: typing.Optional[
+        collections.Mapping[int, items.ItemStatsView]
+    ] = attrs.field()
+    """A mapping of the item instance id to a view of its stats.
+
+    This will be available when `aiobungie.ComponentType.ITEM_STATS` is passed to the request.
+    otherwise will be `None`.
+    """
+
+    sockets: typing.Optional[
+        collections.Mapping[int, collections.Sequence[items.ItemSocket]]
+    ] = attrs.field()
+    """A mapping from the item instance id to a sequence of inserted sockets into it.
+    
+    This will be available when `aiobungie.ComponentType.ITEM_SOCKETS` is passed to the request.
+    otherwise will be `None`.
+    """
+
+    reusable_plugs: typing.Optional[
+        collections.Mapping[int, collections.Sequence[items.PlugItemState]]
+    ] = attrs.field()
+    """If the item supports reusable plugs,
+    this is the mapping from the item instance id to a sequence of plugs that are allowed to be used for the socket.
+    
+    This will be available when `aiobungie.ComponentType.ITEM_SOCKETS` is passed to the request.
+    otherwise will be `None`.
+    """
+
+    plug_objectives: typing.Optional[
+        collections.Mapping[
+            int, collections.Mapping[int, collections.Collection[records_.Objective]]
+        ]
+    ] = attrs.field()
 
     # TODO: Is this needed?
     # talen_grids: ...
-    
-    plug_states: typing.Optional[collections.Collection[profile.ProfilePlug]] = attrs.field()
 
-    item_perks: typing.Optional[collections.Mapping[int, collections.Collection[items.Perk]]] = attrs.field()
+    plug_states: typing.Optional[
+        collections.Sequence[items.PlugItemState]
+    ] = attrs.field()
+    """A sequence of the plug states.
 
-@attrs.define(kw_only=True, weakref_slot=False, slots=False)
+    This will be available when `aiobungie.ComponentType.ITEM_SOCKETS` is passed to the request.
+    otherwise will be `None`.
+    """
+
+    def any(self) -> bool:
+        """Returns `True` if one if the components are available, `False` otherwise."""
+        return any(
+            (
+                self.instances,
+                self.render_data,
+                self.stats,
+                self.sockets,
+                self.reusable_plugs,
+                self.plug_objectives,
+                self.plug_states,
+            )
+        )
+
+    def all(self) -> bool:
+        """Returns `True` if all components are available, `False` otherwise."""
+        return all(
+            (
+                self.instances,
+                self.render_data,
+                self.stats,
+                self.sockets,
+                self.reusable_plugs,
+                self.plug_objectives,
+                self.plug_states,
+            )
+        )
+
+
+@attrs.define(kw_only=True, slots=False)
 class VendorsComponent:
     """Represents vendors-only Bungie component."""
 
     # TODO: Impl this.
 
 
-@attrs.define(kw_only=True, weakref_slot=False, slots=False)
+@attrs.define(kw_only=True, slots=False)
 class StringVariableComponent:
     """Represents the profile string variable component.
 
@@ -235,7 +320,7 @@ class StringVariableComponent:
     """A mapping from the character id to a mapping from an expression mapping definition hash to its value."""
 
 
-@attrs.define(kw_only=True, weakref_slot=False, slots=False)
+@attrs.define(kw_only=True, slots=False)
 class MetricsComponent:
     """Represents the profile metrics component.
 
@@ -261,8 +346,8 @@ class MetricsComponent:
     """The metrics presentation root node hash."""
 
 
-@attrs.define(kw_only=True, weakref_slot=False)
-class CharacterComponent(RecordsComponent, ItemsComponent, VendorsComponent):
+@attrs.define(kw_only=True)
+class CharacterComponent(RecordsComponent, VendorsComponent):
     """Represents a character-only Bungie component.
 
     This includes all components that falls under the character object.
@@ -298,9 +383,7 @@ class CharacterComponent(RecordsComponent, ItemsComponent, VendorsComponent):
     is passed to the request components.
     """
 
-    progressions: typing.Optional[character_.CharacterProgression] = attrs.field(
-        repr=False
-    )
+    progressions: typing.Optional[character_.CharacterProgression] = attrs.field()
     """The character progression component.
 
     Notes
@@ -310,14 +393,14 @@ class CharacterComponent(RecordsComponent, ItemsComponent, VendorsComponent):
     is passed to the request components.
     """
 
-    render_data: typing.Optional[character_.RenderedData] = attrs.field(repr=False)
+    render_data: typing.Optional[character_.RenderedData] = attrs.field()
     """The character rendered data component.
 
     This will always be `None` unless `aiobungie.ComponentType.RENDER_DATA`
     is passed to the request components.
     """
 
-    activities: typing.Optional[activity.CharacterActivity] = attrs.field(repr=False)
+    activities: typing.Optional[activity.CharacterActivity] = attrs.field()
     """A sequence of the character activities component.
 
     This will always be `None` unless `aiobungie.ComponentType.CHARACTER_ACTIVITES`
@@ -333,12 +416,15 @@ class CharacterComponent(RecordsComponent, ItemsComponent, VendorsComponent):
     is passed to the request components.
     """
 
+    item_components: typing.Optional[ItemsComponent] = attrs.field()
+    """A component that includes all items components for this character component."""
 
-@attrs.define(kw_only=True, weakref_slot=False)
+
+@attrs.define(kw_only=True)
 class Component(
     ProfileComponent, RecordsComponent, StringVariableComponent, MetricsComponent
 ):
-    """Concerete implementation of all Bungie profile components.
+    """Concerete implementation of all Bungie components.
 
     This includes all profile components that are available and not private.
 
@@ -380,20 +466,7 @@ class Component(
 
     Included Components
     -------------------
-    - `Profiles`
-    - `ProfileInventories`
-    - `ProfileCurrencies`
-    - `ProfileProgression`
-    - `Characters`
-    - `CharacterInventories`
-    - `CharacterProgression`
-    - `CharacterRenderData`
-    - `CharacterActivities`
-    - `CharacterEquipments`
-    - `StringVariables`
-    - `Records`
-        - `ProfileRecords`
-        - `CharacterRecords`
+    - All
     """
 
     characters: typing.Optional[
@@ -458,4 +531,26 @@ class Component(
     This will always be `None` unless `aiobungie.ComponentType.CHARACTER_EQUIPMENT`
     is passed to the request components.
     """
-    # TODO: Impl other components that don't fit anywhere here.
+
+    transitory: typing.Optional[fireteams.FireteamParty] = attrs.field()
+    """Profile Transitory component.
+
+    This component is used to show minimal information about the player's current fireteam party along
+    with the its members and the activity.
+
+    This will always be `None` unless `aiobungie.ComponentType.TRANSITORY`
+    is passed to the request components.
+    """
+
+    item_components: typing.Optional[ItemsComponent] = attrs.field(repr=False)
+    """A component that includes all items components for this profile component."""
+
+    profile_plugsets: typing.Optional[
+        collections.Sequence[items.ItemSocket]
+    ] = attrs.field()
+    """A sequence of the profile's plug sets."""
+
+    character_plugsets: typing.Optional[
+        collections.Mapping[int, collections.Sequence[items.ItemSocket]]
+    ] = attrs.field()
+    """A mapping from the character's id to a sequence of plug objects bound to that character."""